make the glyph_t typedef 32bit instead of 16bit, in preparation for some changes to the X11 font engine
[git-p4: depot-paths = "//depot/qt/main/": change = 165652]

+++ src/gui/text/qfontengine_p.h
@@ -36,7 +36,7 @@
 
 struct glyph_metrics_t;
 class QChar;
-typedef unsigned short glyph_t;
+typedef unsigned int glyph_t;
 class QOpenType;
 class QPainterPath;
